requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
Django==4.2.9
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0rc1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.24.4
openai==1.61.1
packaging==24.2
pandas==2.2.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
# ✅ Added missing package
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
yt-dlp==2025.1.26
python-telegram-bot==20.0
python-dotenv==1.0.1